Moved turning off SSL certificates to Toolkit class
